Register	account			COMPLETE	COMPLETE             Register	accounts             Register	admin_account             Register	admin_location             Register	forgot_password			COMPLETE	COMPLETE             Register	ldap_test			???	???             Register	login			COMPLETE	COMPLETE             Register	new_customer			COMPLETE	COMPLETE             Register	organization			COMPLETE	COMPLETE             Register	organizations			N/A	N/A             Register	pending_customers		NO	N/A	N/A             Register	privileges			COMPLETE	COMPLETE             Register	register			COMPLETE	COMPLETE             Register	reset_password			COMPLETE	COMPLETE             Register	role			COMPLETE	COMPLETE

+++ modules/register/default/privileges_mc.php
@@ -3,36 +3,40 @@
     $page->requirePrivilege('manage privileges');
 
     if (! empty($_REQUEST['newPrivilege'])) {
-        $privilege = new \Register\Privilege();
-        if ($privilege->add(array('name' => $_REQUEST['newPrivilege']))) {
-            $page->success = "Privilege Added";
-        }
-        else {
-            $page->addError("Error adding privilege: ".$privilege->error());
-        }
-    }
-	elseif ($_REQUEST['btn_update'] == "Update") {
-        $privilege = new \Register\Privilege($_REQUEST['privilege_id']);
-        if ($privilege->id) {
-			if (!$privilege->update(array('name' => $_REQUEST['name']['privilege_id'], 'module' => $_REQUEST['module'][$privilege->id]))) {
-				$page->addError("Error updating privilege: ".$privilege->error());
-			}
-		}
-	}
-    elseif ($_REQUEST['btn_delete'] == "Delete") {
-        $privilege = new \Register\Privilege($_REQUEST['privilege_id']);
-        if ($privilege->id) {
-            if ($privilege->delete()) {
-                $page->success = "Privilege '".$privilege->name."' deleted";
-            }
-            else {
-                $page->addError("Failed to delete privilege: ".$privilege->error());
+        if (! $GLOBALS['_SESSION_']->verifyCSRFToken($_POST['csrfToken'])) {
+            $page->addError("Invalid Request");
+        } else {
+            $privilege = new \Register\Privilege();
+            if ($privilege->add(array('name' => $_REQUEST['newPrivilege']))) {
+                $page->success = "Privilege Added";
+            } else {
+                $page->addError("Error adding privilege: ".$privilege->error());
             }
         }
-		else {
-			$page->addError("Privilege not found");
-		}
+    } elseif ($_REQUEST['btn_update'] == "Update") {
+	        if (! $GLOBALS['_SESSION_']->verifyCSRFToken($_POST['csrfToken'])) {
+            	$page->addError("Invalid Request");
+	        } else {
+                $privilege = new \Register\Privilege($_REQUEST['privilege_id']);
+                if ($privilege->id) {
+			        if (!$privilege->update(array('name' => $_REQUEST['name']['privilege_id'], 'module' => $_REQUEST['module'][$privilege->id]))) $page->addError("Error updating privilege: ".$privilege->error());
+		        }
+	        }
+	} elseif ($_REQUEST['btn_delete'] == "Delete") {
+	    if (! $GLOBALS['_SESSION_']->verifyCSRFToken($_POST['csrfToken'])) {
+        	$page->addError("Invalid Request");
+	    } else {
+            $privilege = new \Register\Privilege($_REQUEST['privilege_id']);
+            if ($privilege->id) {
+                if ($privilege->delete()) {
+                    $page->success = "Privilege '".$privilege->name."' deleted";
+                } else {
+                    $page->addError("Failed to delete privilege: ".$privilege->error());
+                }
+            } else {
+			    $page->addError("Privilege not found");
+		    }
+	    }
     }
     $privilegeList = new \Register\PrivilegeList();
     $privileges = $privilegeList->find(array('_sort' => 'module'));
-?>
